refactor: consolidate analytics types
