+++ packages/@ourworldindata/types/src/analyticsTypes/analyticsTypes.ts
@@ -1,4 +1,4 @@
-// Analytics Event Parameter Type System
+// Analytics Type System
 //
 // This file defines the complete contract between the frontend and analytics backend.
 // Each event category has exactly one corresponding parameter interface.
@@ -19,12 +19,69 @@
 //
 // 3. This file is the single source of truth for the analytics contract
 
-import {
-    GrapherImageDownloadEvent,
-    GrapherInteractionEvent,
-    EntitySelectorEvent,
-    EntityControlEvent,
-} from "./events.js"
+// =============================================================================
+// EVENT CATEGORIES
+// =============================================================================
+
+export enum EventCategory {
+    Filter = "owid.filter",
+    GlobalEntitySelectorUsage = "owid.global_entity_selector_usage",
+    GrapherView = "owid.grapher_view",
+    GrapherClick = "owid.grapher_click",
+    GrapherHover = "owid.grapher_hover",
+    GrapherError = "owid.grapher_error",
+    GrapherEntitySelector = "owid.grapher_entity_selector",
+    ExplorerView = "owid.explorer_view",
+    Expander = "owid.expander",
+    ExplorerCountrySelector = "owid.explorer_country_selector",
+    KeyboardShortcut = "owid.keyboard_shortcut",
+    SiteClick = "owid.site_click",
+    SiteError = "owid.site_error",
+    SiteInstantSearchClick = "owid.site_instantsearch_click",
+    SiteSearch = "owid.site_search",
+    SiteSearchAutocompleteClick = "owid.site_search_autocomplete_click",
+    SiteSearchResultClick = "owid.site_search_result_click",
+    SiteFormSubmit = "owid.site_form_submit",
+    DetailOnDemand = "owid.detail_on_demand",
+    SiteGuidedChartLinkClick = "owid.site_guided_chart_link_click",
+    SiteChartPreviewMouseover = "owid.site_chart_preview_mouseover",
+}
+
+// =============================================================================
+// EVENT-TO-PARAMETER MAPPING
+// =============================================================================
+
+export type EventParamsMap = {
+    [EventCategory.SiteSearch]: SiteSearchParams
+    [EventCategory.SiteSearchAutocompleteClick]: SiteSearchAutocompleteClickParams
+    [EventCategory.SiteSearchResultClick]: SiteSearchResultClickParams
+    [EventCategory.GrapherView]: GrapherViewParams
+    [EventCategory.GrapherClick]: GrapherClickParams
+    [EventCategory.GrapherHover]: GrapherHoverParams
+    [EventCategory.GrapherError]: GrapherErrorParams
+    [EventCategory.GrapherEntitySelector]: GrapherEntitySelectorParams
+    [EventCategory.ExplorerView]: ExplorerViewParams
+    [EventCategory.ExplorerCountrySelector]: ExplorerCountrySelectorParams
+    [EventCategory.DetailOnDemand]: DetailOnDemandParams
+    [EventCategory.Expander]: ExpanderParams
+    [EventCategory.KeyboardShortcut]: KeyboardShortcutParams
+    [EventCategory.SiteGuidedChartLinkClick]: SiteGuidedChartLinkClickParams
+    [EventCategory.SiteChartPreviewMouseover]: SiteChartPreviewMouseoverParams
+    [EventCategory.SiteClick]: SiteClickParams
+    [EventCategory.SiteFormSubmit]: SiteFormSubmitParams
+    [EventCategory.SiteInstantSearchClick]: SiteInstantSearchClickParams
+    [EventCategory.SiteError]: SiteErrorParams
+    [EventCategory.Filter]: FilterParams
+    [EventCategory.GlobalEntitySelectorUsage]: GlobalEntitySelectorUsageParams
+}
+
+export type GAEvent = {
+    [K in EventCategory]: { event: K } & EventParamsMap[K]
+}[EventCategory]
+
+// =============================================================================
+// EVENT PARAMETER INTERFACES
+// =============================================================================
 
 // Site Events
 
@@ -251,4 +308,38 @@ export interface GlobalEntitySelectorUsageParams {
     eventAction: EntityControlEvent
     /** Additional note or context */
     eventContext?: string
+}
+
+// =============================================================================
+// EVENT ACTION TYPES & HELPERS
+// =============================================================================
+
+export type EntityControlEvent = "open" | "change" | "close"
+
+export type EntitySelectorEvent =
+    | "enter"
+    | "select"
+    | "deselect"
+    | "sortBy"
+    | "sortOrder"
+    | "filterBy"
+    | "clear"
+
+export type GrapherImageDownloadEvent =
+    | "chart_download_png"
+    | "chart_download_svg"
+
+export type GrapherInteractionEvent =
+    | "map_country_hover" // hover over a country in the map or on the globe
+    | "map_legend_hover" // hover over a legend item on the map tab
+
+export enum GrapherErrorAction {
+    GrapherViewError = "grapher_view_error",
+    EntitiesNotFound = "entities_not_found",
+}
+
+export interface GrapherAnalyticsContext {
+    slug?: string
+    viewConfigId?: string
+    narrativeChartName?: string
 }
